backend/requirements.txt
fastapi[all]
sqlalchemy
aiosqlite
openai
python-dotenv
alembic
uvicorn
